LayoutTests/fast/canvas/CanvasRenderingContext2DSettings-pixelFormat-disabled-expected.txt

Test that the pixelFormat member of CanvasRenderingContext2DSettings is unsupported when CanvasPixelFormatEnabled=false.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testSettings.pixelFormat is undefined.
PASS testSettings.pixelFormat is undefined.
PASS testSettings.pixelFormat is undefined.
PASS testSettings.pixelFormat is undefined.
PASS successfullyParsed is true

TEST COMPLETE
